+++ bartons/src/lib.rs
@@ -21,6 +21,7 @@ fn python_kernels(_py: Python, m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_function(wrap_pyfunction!(kernels::trange::trange_py, m)?)?;
     m.add_function(wrap_pyfunction!(kernels::atr::atr_py, m)?)?;
     m.add_function(wrap_pyfunction!(kernels::mad::mad_py, m)?)?;
+    m.add_function(wrap_pyfunction!(kernels::cci::cci_py, m)?)?;
     m.add_function(wrap_pyfunction!(samples::random_prices_py, m)?)?;
     m.add_function(wrap_pyfunction!(samples::with_n_chunks_py, m)?)?;
     Ok(())